Changed Code to match coding Standards Look in description for details
Fixed a typo with DBEmployee -> EmployeeDB
Changed EmployeeController match EmployeeDBIF
Changed a wrong Attribute in EmployeeDBIF to match earlier documentation
Updated findEmployeeByEmployeeId to match the changed attribute
Updated use of doubles to be BigDecimal
Added Null checks in OrderController
Removed Redundencies in LogController

+++ src/model/HourLog.java
@@ -1,18 +1,20 @@
 package model;
 
+import java.math.BigDecimal;
+
 public class HourLog extends Logs {
 
-	private double hoursWorked;
+	private BigDecimal hoursWorked;
 
-	public HourLog(Employee employee, double hours) {
+	public HourLog(Employee employee, BigDecimal hours) {
 		super(employee);
 		hoursWorked = hours;
 	}
 	
-	public double getHoursWorked() {
+	public BigDecimal getHoursWorked() {
 		return hoursWorked;
 	}
-	public void setHoursWorked(double hoursWorked) {
+	public void setHoursWorked(BigDecimal hoursWorked) {
 		this.hoursWorked = hoursWorked;
 	}
